requirements.txt
quart
quart-cors
lxml
Pillow
python-pptx
boto3
